Exclude /mnt from systemd-logind restrictions on Bullseye and later
Starting with Bullseye the systemd unit for systemd-logind uses ProtectSystem=strict,
which doesn't work with HDFS and results in a failing systemd-logind service.

Fix this by excluding /mnt from the list of inaccessible paths for
the systemd-logind service (shipped via an override file to the systemd
unit). This will also affect the Hadoop cluster once it moves to Bullseye.

Hosts: clouddumps1001.wikimedia.org
Bug: T316123
Change-Id: Id9e6a8d5e059b6e0f1d2f041cf1d913870b94b09