role::ci::master: remove admin dependency hack
Since I508229fbe we're using the admin module to manage the docker group
membership for the jenkins user, but this hack was not removed.

Now with the introduction of user dependencies on shell packages for the
creation of users in I7af14e3449988 that caused a circular dependency
on all contint hosts in production, making puppet fail.

Change-Id: I7241301733e8e3c729dc255a463d61b13e9b3276