bacula: Add new hosts backup1009 & backup2009 as new storage servers
backup1009 and backup2009 will soon be scheduled to contain regular
production backups of server's filesystems to increase available
capacity.

For now, set them up as additional storage servers available from
backup1001, but with no configuration or schedule on the director.

This finishes the split of the old ::backup role into
backup::databases and backup::production, although the old profile
will be kept while data can be recovered from it.

Bug: T313582
Change-Id: I9dedc2c4b9c992f9da7479c46e60bbcd67f8b082