phabricator: ensure only the one active_server connects to rw mysql
Only the single server set as $active_server should write to
databases.

All other servers should only use the slave which is read-only.

This should be safer and prevent that any additional host during
migrations starts writing to the database.

The puppet role can be applied safely and later the active_server
can be switched.

Bug: T315713
Change-Id: If3c22246d9cdd0abcf7fd7f6424dc4c9f86b290d