gitlab: reduce backup_keep_time to 1d
This change reduces backup_keep_time from 2d to 1d.
This will affect backups on GitLab host, backups in
bacula will be stored longer than 1 day.
This change is not needed immediately, current disk
usage is 25% for backups. But this change was discussed
in T274463 as the long-term plan. So long-term only the
most recent backup is stored on the production
instance and all other backups in bacula.

Bug: T274463
Change-Id: I7c9afcff2a035fbf9e1ee258c0c585531f1487a8