Add alert for CirrusSearch reported memory issues
New alert triggers based on the number of requests that trigger the
circuit_breaking_exception in elasticsearch. These errors occur when
elasticsearch estimates the request would push jvm heap usage beyond
a predefined limit.

Current alerting values are a bit of a guess. We expect these errors
to be quite rare except in cases where one (or more) of the search
instances are having memory pressure issues.

Bug: T316712
Change-Id: I3a8191be93a6ac1cc62cc80aa7a7f97baae417ab