phabricator: Include scap "done" rev in pre-finalize permissions reset
Stop excluding the "done" rev directory as it changes between the time
the promote stage check runs and the clean occurs during the finalize
stage. In order words, the "done" rev at this point will be considered
for clean up during finalize and must have the right ownership/mode for
directory removal to succeed.

Bug: T313953
Change-Id: I2580a9c7e48c0470d01719f003c05ec32b6a9ad2