phabricator: ensure home dir exists before sysuser is created
When applying the role on a new host and now using systemd::sysuser
to create the phd user, we are getting:

Systemd::Sysuser[phd]/User[phd]: Skipping because of failed dependencies

and the /var/run/phd dir does not exist yet.

Add a require attempting to fix that and turn the home dir into a class
parameter along the other existing directories.

Bug: T280597
Change-Id: I5f8d660e10c37f40506ca0ab36004abe4231068e