phabricator::phd: actually use $phd_user variable and small improvements
Actually use the $phd_user variable everywhere instead of hardcoding
it in half the places while also having it as a parameter.

Same for the home dir added in I5f8d660e10c37f.

Also move the UID number to a parameter with the current reserved
value as default.

Bug: T280597
Change-Id: I7c5e0605df42195e8884361bf43cfc00e640bbe3