Grant ttaylor access to PII through Superset
The previous changed added ttaylor to the ldap_only_users structure,
which is insufficient to allow access to Superset and PII data. This
change moves the ttaylor account into the users structure, but with an
empty ssh key.

In addition to moving the account, it is also added to the
analytics-privatedata-users group.

Bug: T292299
Change-Id: Id3c28b28624a80f1ba36f8072ce43c89a42adb9a