Use the dumpsgen user to mount the NFS dumps directories
We need to be able to set the ownership of the /mnt/nfs/dumps-*
directories to be dumpsgen:dumpsgen wherever it is mounted. Currently
puppet is failing to set the ownership to root:root when it is mounted.

This change applies the dumpsgen user and group as part of the
statistics::dataset_mount profile and then uses it to set the ownership
of the mount points.

Hosts: P:statistics::dataset_mount
Bug: T317359
Change-Id: I3e61499f007ba4f75ce1628298e6d2458cdb2e99