Move kafka on kafka-logging2001 to PKI TLS certificates
This is the first Kafka broker to move to the new TLS certificates
generated by the PKI infrastructure (as opposed to Puppet's CA).

All the brokers are already using a CA bundle able to validate
TLS certificates generated by the Puppet CA and by PKI, so
we can move one broker at the time to the new TLS certs
monitoring how things go.

Bug: T300130
Change-Id: Id0252578cfffec21b461f6cdd8500ee372e2c9c2