openldap: Include slapd-audit.log to backup
This file contains the audit log of slapd. We have currently no real tooling
to parse or rotate this file, but it's needed from time to time to understand
when a given change was made and by whom.

Current capacity is rather small (1G on serpens and 2G on seaborgium; and
this is the accumulated data over like six years).

Bug: T317516
Change-Id: Ifbec5be0dd4d4af8e7b51fa90afaad0bd452daec