mail::mx: Modify the Received header
Prior to this change emails that are sent via authenticated relay to the
foundation are flagged as SPF softfail by google. This is due to how we
relay emails to google.

[1] smtp auth -> [2] mx1001.wikimedia.org -> [3] google

This flow produces the following header:

Received: from [104.131.61.189] (port=56008 helo=mbuki-mvuki)
  by mx1001.wikimedia.org with esmtpsa
  (TLS1.3) tls TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384
  (Exim 4.94.2)
  (envelope-from <qualtrics@wikimedia.org>)
  id 1oXqAN-001aL2-OT
  for jhathaway@wikimedia.org; Mon, 12 Sep 2022 20:36:27 +0000

Even though the [2] hop is allowed to send as wikimedia.org, Gmail flags
the mail as a SOFTFAIL because hop [1] is not allowed to send as
wikimedia.org according to our SPF record.

After this change we hide the ip and show the authenticated_id instead:

Received: from qualtrics@wikimedia.org
  by mx1001.wikimedia.org with esmtpsa
  (TLS1.3) tls TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384
  (Exim 4.94.2)
  (envelope-from <qualtrics@wikimedia.org>)
  id 1oXqFm-001er1-Lt for
  jhathaway@wikimedia.org; Mon, 12 Sep 2022 20:42:02 +0000

Hosts: O:mail::mx
Bug: T317574
Change-Id: I4cf743f97438361a4c13ce9cb02aa28ec86e76dd