phabricator: Allow deploy user to keep scap3 environment variables with sudo
This is copying I6a28af469c62e98728 by Dan Duvall just keeping
it limited to Phabricator instead of changing scap::target globally
right away (as it was in the first patch sets over there as well).

This way we can unblock Phabricator deployment and test it without
risking to affect mw appservers and others and other reviewers
can review the global change.

Bug: T313259
Change-Id: Ie44a16e6b6d883d67279f23ddf7adba7509c05b7