gerrit: change its templates to regular files
The its-phabricator plugin relies on soy template to generate the
message emitted to Phabricator. They used to be regular files but
converted to Puppet templates with c791098726a in order to craft shorter
URLs:

  {$formatChangeUrl}
  https://<%= @host %>/r/{$changeNumber}

@host is the sole variable used in those templates, I don't think there
is any need to make it variant given we have a single Gerrit in
production.

While at it have Puppet manage the template directory with a single
Puppet file statement.

Change-Id: Id01a382a565172c869a00e4f5d69da8054c14cd9