dbbackups: Increase backup freshness check from 8 to 9 days
Because dumps for es* hosts can take more than 1 day to run, and
the "freshness" of a backup is compared with the start of the
previous backup (when the backup is consistent to), more than 8
days can pass between a backup and the completion of the one
the next week.

Amend the expected freshness from 8 days to 8, in seconds.

Change-Id: Id1d67b34b04e8e22249e75cba0447bb86b61dcc3