phabricator: remove Icinga monitoring for phd supervising processes
This alert has created false positives because the number of supervised
processes can change.

In the ticket discussion we agreed that we really just care that phd
is running at all and not how many processes run by user phd.

And given the "move stuff away from Icinga" mission we should just
remove this.

It's more relevant anyways to just monitor if Phabricator is actually
up for the end user and I don't recall this finding an actual problem
for us.

Bug: T315962
Change-Id: Iab3aa58eb2287c60568c8252169091ad595d559b