gerrit: rm redundant service_params ensure => running
Our systemd::service classes has a default set of parameters passed to
the service:

  $base_params = {
      ensure   => stdlib::ensure($ensure, 'service'),
      enable   => $enable,
      provider => $provider
  }

Thus `systemd::service { ensure => present }` would result in the
service to have `ensure => running` and we do not need to pass it
explicitly via `service_params`.

Remove the redundant parameter and add spec to cover the service is
ensured as running.

Change-Id: I5a069a937e7063fe32071d36e1a9ef757e8e06ec