keystone: refactor fernet key rotation timing
Previously this relied on hardcoded rotation times from hiera;
it also had the timing exactly backwards so that new
keys were immediately overwritten via an rsync call.

The new plan uses division to distribute key rotations
around the clock, and organizes the rsyncs so that
each change is propogated to the other keystone hosts
after a key rotation.

In addition to removing some hiera, this has the bonus
advantage of preventing simultaneous rsyncs to any given
host which should avoid collision errors.

Bug: T317838
Change-Id: Ifc5b44f5b9683ded3568070ddab70b719e3c7bb6