buildkitd: Support configuration of OCI executor nameservers
When buildkitd's OCI executor spawns a container, it looks at the host
`/etc/resolv.conf` for non-loopback entries (127.0.0.0/8) and if it
can't find at least one, it defaults to Google public nameservers
(4.4.4.4, 8.8.8.8, etc.). Since we're running buildkitd in a container
using a custom Docker network, the host nameserver will be `127.0.0.11`
(Docker's DNS proxy on a loopback address) and the mentioned defaults
will be set. We don't want this. We want our internal nameservers to be
used instead.

Buildkit supports (though it's undocumented) configuration of explicit
DNS servers to be used by its executors. Let's allow nameservers to be
configured via `profile::gitlab::runner::buildkitd_nameservers` and
default the value to `profile::resolving::nameservers` which should be
defined for all nodes/roles using `profile::base`.

Bug: T317904
Change-Id: Ic985fea2adddb831e7df3d1bd9f12fd4e9fd1298