debconf::set - add $owner param, set owner in conda_analytics/init.pp
The first element in the debconf-set-selections command should be the
'owner' of the setting, which is usually the package name.
debconf::set currently uses 'set' as the owner for all its usages.
Adding this $owner param allows us to set the correct package name owner
for debconf settings.

For historical reasons, we default $owner to 'set', to make this change
a no-op for existing usages.

Bug: T312882
Change-Id: Ie08418c47d156dd1922144c153963e84325a11af