Deploy Spark 3 on the whole production cluster
The whole install is currently deployed on:
* statistics/explorer machines
* an-worker1111

We deployed it to an-worker1111 to check if /opt/conda-analytics/pkgs
dir has been removed in deb postinst script as it should. And it
works.

We didn't go further because the installation on the whole worker pool will
take time, should be monitored, and reverted in case of problems. Better
doing this operation on Monday.

Next steps:
* merge this pull request
* wait for puppet to run on the whole cluster
* Pause Airflow analytics (through the UI)
* merge and deploy airflow-dags
  https://gitlab.wikimedia.org/repos/data-engineering/airflow-dags/-/merge_requests/141
* Restart the Airflow systemd (an-launcher1002)
* Unpause Airflow analytics (through the UI)

Bug: T312882
Change-Id: I1770405cf72bfa87a4bfe6ea495e06d214baac11