gitlab_runner: enable unprivileged_userns_clone in WMCS
buildkitd is deployed to WMCS runners. Service start fails with

/proc/sys/kernel/unprivileged_userns_clone needs to be set to 1.

This change enables unprivileged_userns as a fix. The error
can also be fixed by disabling buildkitd for shared runners.
I'm not sure if there sould be buildkitd support on shared runners.

If buildkitd is needed on WMCS Runners, remember that CI jobs
use the default network and buildkitd is running in
gitlab-runner bridge network (in WMCS).

Bug: T307810
Bug: T308271
Change-Id: Ifb8196d0418eb526d45321b027e659d84b2b3b23