C:memcached Fix memcached bootstrap
On bootstrap, override is deployed, systemd daemon-reload is done, but
we don't restart the memcached service, resulting in a non-functional
memcached host on bootstrap without manual service restart. It is not
desirable to restart memcached automatically on config file change.
On the same pattern as I37c421e9a using systemctl mask/unmask, we make
sure that memcached is not started with the default configuration, and
keep control of restart on config file change.

Bug: T318697
Hosts: mc1037.eqiad.wmnet, mc2020.codfw.wmnet
Change-Id: Ieec8714be8b420c2b57cabaa11c76805d874cf3a