cache node disk layout p11n for F4 config
This adapts both the netboot partman selection and the late
command nvme setup to cover existing cases and the new cp4/5 nodes
that we'll be installing soon.

In the netboot case, we're switching to explicitly calling out the
Legacy (soon to be decommed as they're replaced!) cases and then
covering all the modern ones under a simpler regex for all cp
nodes at all known sites.

In the late_command case, it's simpler to do it by iterating the
available nvme devices and make it universal (for the cache
nodes).  This block will do nothing on the remaining Legacy hosts,
and will iterate the 1-2 drive cases for all the others.

https://wikitech.wikimedia.org/wiki/Traffic_cache_hardware for
context on the new hardware that's rolling out.

Bug: T317244
Change-Id: I59da036925b58bec5d71439e9788f37b076613f8