keyholder: Switch shebang to /usr/bin/python3
No need for env since we only ever want to use the distro-provided Python
and not using env makes keyholder show up as the process name in lsof
(which is what we parse to detect necessary service restarts after
library updates), otherwise it simply shows as as python3.

Also axe the full Apache license statement, since we now have a proper
SPDX header.

Change-Id: Ifde09fda747bf38aa328ebd9bffd2b56db7af53c