openstack: neutron: l3_agent: don't sysctl base interface
The current setup for the neutron l3 agent uses VLAN-tagged sub-interfaces for the
dataplane.

But we were applying sysctl also for the base interface and this might be problematic if
the base interface is mean to hold both dataplane and controlplane, which is what we are
trying to do to reduce NIC footprint in the datacenter.

The solution might be to simply don't apply sysctl config the base interface, because it is
now both dataplane/controlplane. Only VLAN-tagged sub-interfaces are dataplane.

Anyway, settings to the base interface should (in theory) not interfere with settings
for sub-interfaces.

Bug: T318824
Bug: T277287
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Change-Id: I3a5263535f904660f35700a052a80f5e437a1ef3