openstack: neutron: introduce workaround for debian bug #989162
The debian bug #989162 https://bugs.debian.org/989162 in the bridge-utils package
is preventing us from having a single NIC for both dataplane and controlplane in
neutron-enabled openstack servers.

What happens is:
 * a base interface (eno1) is configured as a standard interface with IPv4/IPv6 addresses
 * vlan sub-interfaces (eno1.xxxx and eno1.yyyy) are created and added to the neutron bridge
 * the bridge-utils code disables IPv6 on the base interface (eno1) via ifup hook
 * ifup fails to configure eno1 because we try to add IPv6 to a disabled_ipv6 interface

This patch introduces the required patch for the bridge-utils package, but only for the
codfw1dev deployment, until we make sure is perfectly safe to deploy on eqiad1 as well.

I decided to go with a separate module (bridge_utils) because the same workaround could
be useful perhaps in other virtual environments (ie, ganeti).

Bug: T318824
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Change-Id: I74c57d49dc7790bee11cd6cb5d7de8c1d9e2b969