mariadb: Set binlog format for dbstore mariadb databases to ROW
The fear is that even if binary log is not enabled (explicitly
setup on this patch, but unchanged), by default the binlog format
is MIXED, which could allow some extra locking, impacting
reimports.

This was proven not to fix the issue at T319383, but still is
probably a good idea to set it to ROW, as core,
even if no binary logs are available, for consistency with
production.

Bug: T318062
Change-Id: I6d278efdc6f265edc29645c7c2a9f13829b40bd2