cloudnet1005/1006: prepare for single NIC setup
We are trying to introduce single NIC setup for all cloud servers to reduce
switch port footprint in the DC.

Since we have cloudnet1003/1004 in service and we don't want to modify them,
introduce host-level hiera overrides to prepare the new servers (1005/1006)
before the migration.

The original NIC name in these 2 is called 'enp175s0f0np0'. That string,
plus the VLAN tag (like 'enp175s0f0np0.xxxx') fails the IFNAMSIZ check,
so we have to rename the interfaces as well.

The string `nic0` is an arbitrary string, explicitly not using `eth0`
to make it clear is a made up name.

Bug: T316284
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Change-Id: I628f09146a3123bb9b330e521d14715ff8efda7d