role::kafka::logging: final clean up after migrating to PKI
Remove the settings that allowed the migration to PKI.
Since we are using new TLS certificates with the kafka hostname
as CN, we don't need anymore to have the old CN "kafka-logging"
and it should be removed. The update requires a run of the
kafka roll restart cookbook on both clusters.

Bug: T300130
Change-Id: I240e5a5f94e1f8ff52b0855736572d8a8e480315