cumin2002: Add an hourly httpbb run against mw2271
Now that we're serving out of both eqiad and codfw, let's alert on
httpbb failures in either site. Most of the time, this will be redundant
(both OK or both failing) but if it ever isn't, we'll want to know that.

Testing from both cumin hosts also ensures we'll still have visibility
when one or the other is out of service.

Change-Id: I8213069f268f5aa40a3c6aae7a88c4e2025a3037