vrts: allow installing a local mariadb server in cloud
To run VRTS for testing purposes in cloud VPS, allow
installing a local database server.

Use the same include used in existing role(simplelamp2)
but without the other services that come with it (apache, memcached).

Add a parameter to toggle it in Hiera and set the default to false
so that production stays unchanged.

Bug: T317059
Change-Id: I45fe12ad491fedf8065d99b5fa7ce16b15097f08