openstack: neutron: refresh bridge ifupdown code to handle ordering
There could be cases in which the bridge ports are not online. Then ifupdown
will create the bridge device but it wont contain the port. And that's the
recipe for a network outage.

Try mitigating this scenario by forcing the dependant port interface to be up
in a recursive call to ifup (which is allowed [0]).

This only affects cloudnet nodes, since bridges on cloudvirts are dinamically
populated using other mechanisms (nova-compute, neutron-linuxbridge-agent and friends).

[0] See `CONCURRENCY AND PARALLEL EXECUTION` in interfaces(5)

Bug: T319524
Signed-off-by: Arturo Borrero Gonzalez <aborrero@wikimedia.org>
Change-Id: I115ebd5739b42e99e123b29c4fbb17ca1bf60420