gerrit: use 2 threads to replicate to GitHub
We do a full replication on Gerrit restart but only have one thread when
replicating to GitHub. It took almost 5 hours for a full replication to
complete which today caused a delay for the release of a composer
package (gerrit -> github -> packagist).

Add a second replication thread to speed up the replication.

This does not need Gerrit to be restarted. It is only needed when
restarting Gerrit which will happen at some point in the future anyway.

Hosts: gerrit1001.wikimedia.org
Change-Id: Ieb31dd31b6253642dea132c378f3c17695f1a98e