Switch profile::base::linux510 to the new meta package
Now that Buster is an LTS release, buster-backports is no longer being
updated. The 5.10 kernel packages continue to be updated under a
separate source package name, though: src:linux-5.10 (which ships
rebuilds of what is in Bullseye for Buster).

Switch away from the previous apt pin to the new meta package.

Bug: T319067
Change-Id: I0678269cc6e0d7bbc3025b8b2d9ddcc9ac220856