rsyslog::conf remove trailing newline logic
This logic is currently broken, because the regexp always matches
the first line of the content, regardless of whether it has a trailing
newline. The `$` anchor matches the end of a *line* without the
newline, so a `\n` is always inserted.

Since rsyslogd doesn't mind if files don't have trailing newlines, just
drop the logic rather than reworking it.

Bug: T320569
Hosts: C:rsyslog
Change-Id: I631d13f9cb47a9c80050b0924db59d26cdab6c2c