gitlab_runner: add webproxy to allowed_services
This change adds webproxies in eqiad and codfw to allowed_services.
Technically the access is already working, because webproxy has a
public address and is not inside private network. However to make
sure we track what services are allowed this completes the list.
In case of a network change (move from public to private network)
the webproxy access still works.

Bug: T295481
Change-Id: I2ae83c78d8f749bf6bac03fa3735706e4a8b7d9d