logstash: set webrequest index replicas to 0 for large indexes
The logstash-webrequest indexes are large enough to prevent indexes from
rotating onto the hdd nodes due to space constraints.  This is an
attempt to retain the indexes, but reduce their space requirements by
eliminating replicas of the data.  The loss of this data is less
concerning than the risk of loss for other log data.

Bug: T313099
Change-Id: I422d09bdd19c01f31d4da0f4b3dc14734be85000