phabricator: move list of dumps rsync clients to parameter and Hiera
Move the list of allowed clients to rsync the phab dumps file
out of the middle of the profile and to a proper parameter and Hiera.

No hardcoded values like host names in code.

Also rename to "dumps_rsync_clients" because there are other
types of rsync clients for repos as well and this isn't clear
immediately.

Bug: T313360
Change-Id: I72c7d3bd3b713cb8b8a902951cf9d129b3c1c724