backup: Increase the maximum amount of volumes of esrw (backup*003)
The space used by Read-write External Storage database backups
has increased near the 100TB, create at least 200 volumes of
500GB each so they don't get blocked waiting for expiration on
each backup run.

This happens because we have increased the available space, and
the retention to 90 days, but we hadn't increased the max volume
count.

This will increase the amount of available space for es-rw
backups on both datacenters (backup1003, backup2003) to 100TB.

Bug: T313582
Change-Id: If8973af6d2c91d850b6633be776e4988d0ee4603