Add postgresql to an-db100[1-2]
We have decided to use an-db100[1-2] as a postgresql primary and standby
server, respectively. The initial database to be used on this cluster
will be airflow.

This commit not only adds the postgresql service, but also configures
replication from an-db1001 to an-db1002 as well as backups with bacula.

It takes inspiration from the netbox::db profile, which also configures
postgresql in a similar way.

Note that it does not create the airflow database nor its user accounts.
That will be added in a separate commit.

Bug: T319440
Change-Id: Ie880958c84b8bec69e2350a7b4f61ca058f92e4b