package_builder: add deb-src for buster and bookworm
For building packages, we are currently shipping deb-src entries for
older distros such as stretch. But with buster as the current oldstable
and bookworm as the testing, we should add an entry for that as well so
that if we were to build packages for bullseye using packages built for
buster, we can fetch their source (of the buster package.)

A manual solution involves editing sources.list temporarily and adding
oldstable but that is not desirable.

Bug: T321309
Hosts: O:builder
Change-Id: Icd6ff6babf7376aa5b8dd0c4d95f209d29cdb56b