Add a simple mechanism for creating postgresql users and databases
Our initial requirements for postgresql databases for airflow are simple,
in that we have a 1:1 database to username mapping and no special
privileges are required.

Given that situation, we can look up a list of database names from
hiera and create a postgresql database for each. Similarly, we can look
up an array of database usernames and password from the private repo and
create grants accordingly.

Additional functionality is available from the postgres module, but for
the initial set of databases that we need this simple apprach is fine.

Bug: T319440
Change-Id: I838345ca5ca70b5daa886a9cebe24f4e651d4241