openstack: encapi: perform git updates server-side
To open up the API to third-party clients (Terraform etc.), the
cloud/instance-puppet.git updates need to happen server-side instead of
it being a client responsibility like it is now.

The approach taken here is to use two new MariaDB tables to store the
data to commit the changes to the repository. This has a couple of
benefits over doing the updates during the web requests, namely making
the actual web request processing faster and avoiding issues with
multiple pushes happening at the same time (this is T306642).

As with the dynamicproxy DNS handling rollout, we take an approach where
during the rollout clients can use a special HTTP header to indicate if
they are taking care of the Git update or if that's left to the API
service.

Bug: T306642
Bug: T318504
Depends-On: Ie8d626c19bc64d9e4979a2825cabd9b48242a6f7
Depends-On: Ia5dbd1427641ea38fb1ec126d37458de4f44b5d7
Change-Id: Ibfb80b9dae072578c2a3858264308449307b5227