single_backend mode for production varnishes
This refactors "single_backend_experiment" into the new option
"single_backend", which we'll be applying in production only on
our latest-generation cache nodes as they're deployed.  These new
nodes have greatly increased storage (4-8x, depending on cluster)
in order to offset the downsides observed in the original
experiments.

The setting can be initially applied per-host while we roll new
nodes into a given datacenter, and then later the hieradata can be
moved to the per-DC level once a DC's transition is complete.

It will take a period of years for us to warranty replace all
global cache nodes to increase their capacity, so we'll be
transitionally operating both single and multi -backend
datacenters for quite a while!

The key changes here are:
* The value is now a per-host (or broader) boolean, rather than a
  hostname.
* The option only affects ats-be routing for the host(s) it is
  applied to - other legacy multi-backend nodes in the same DC
  will continue chashing traffic to all available nodes, including
  new ones marked as single-backend.

Bug: T288106
Change-Id: Idb80473c214a5993fa3dfc3f81a304f3cc81b4d7