Remove confd_experiment_fqdn support
In the new iteration of how we apply the "single backend" concept,
we won't be excluding the singular backends from the hash set for
other nodes in the same DC.

We only want the "single backend" concept to impact the individual
nodes it's applied to, while remaining multi-backend nodes
continue to hash over the entire node set in a datacenter.  This
will help them not suffer a deterioriation in hitrate when a
datacenter is partially converted to "single backend" style of
operation.

Bug: T288106
Change-Id: Ie8fa0d7a1da4a698cd52acd8c4771a76d0ff5125