apt::repository: use signed-by instead of apt-key
Updates the apt module to use signed-by instead of apt-key, since
apt-key is deprecated and will start printing deprecation warnings soon.

This is attempt 2 of this change, the previous one was at
Id7167c8e87615d52c75f8fe1dfddfdb0c0477d93. The issue with the previous
one had to do with file permissions; turns out apt runs some operations
as the _apt system user, which previously did not have access to the key
file. This patch changes the permissions from 0400 to 0444 which fixes
the issue.

Signed-off-by: Taavi Väänänen <hi@taavi.wtf>
Change-Id: I558c7ecf38d393f7c5b139e1f717e52f9a74ef30