kubernetes: Rename mwdebug to mw-debug
hieradata: Rename mwdebug to mw-debug in kubernetes mw_releases

deploy-mwdebug.py: Switch deployment to mw-debug
As part of the naming standardization for mediawiki on kubernetes,
the new helm service deployment name is mw-debug. Update hardcoded
values in the deployment script to reflect this change. This needs
to be deployed after the new service is in place.

Bug: T321201
Depends-On: I99e62a75b69b308438a73e90006476506a229204
Change-Id: I277c933fb70772a4eeaaec01fc522a7b33edcf5d