dumps: add sister projects to kiwix dumps rsync
Instead of only downloading wikipedia, add all the sister
projects as well, as requested in the linked ticket.

Since "wikipedia" was hardcoded, added an array to define
all projects and put a for-loop around the existing "do_rsync".

Just not sure if we need to create the initial project
directories first. It seems like we have to.

Bug: T57503
Change-Id: I75861e6e6051795d49641d7610b8816666fc1535