gitlab_runner: enable restrict_firewall for Shared Runners
This change enables restrict_firewall for Shared Runners in WMCS.
Similar to the firewall for Trusted Runners, only access to the
internal wmf network is blocked and needs to be explitly added to
profile::gitlab::runner::allowed_services.
Traffic to public destinations (internet) is not affected by this
change.
It was discussed that traffic from gitlab-runner CI jobs should
not be able to contact other WMCS services by default.

This change also adds NAT rules for the default docker network
172.17.0.0/16. That is needed to ensure containers on Shared
Runners can access external resources.

Bug: T317341
Change-Id: I1f1a112b37d308ffa10ab8a95205815250c9b7ec