opensearch: make upgrade-phatality.sh stricter
Both versions of our script would continue installing Phatality even if
the removal failed. Doing a deployment I have encountered:

 Unable to remove plugin because of error:
 "ENOTEMPTY: directory not empty,
  rmdir '/usr/share/opensearch-dashboards/plugins/phatality'"

The script ignores the error and process with the install which fails
due to the plugin already existing:

 Attempting to transfer from
   file:///srv/deployment/releng/phatality/deploy/phatality-2.2.0.zip
 Transferring 13336 bytes....................
 Transfer complete
 Retrieving metadata from plugin archive
 Extracting plugin archive
 Extraction complete
 Plugin phatality already exists, please remove before installing a new version

Only attempt to remove the Phatality plugin when it is installed. This
is done by looking it up in the output of `kibana-plugin install`.

Instruct bash to abort on script errors (-e), unset variable (-u) and
also when a command fails in a pipe chain (which is not used currently
but might later on and is a good practice to make bash stricter).

Also double quote a variable to please shellcheck.

Bug: T304440
Change-Id: I25a956c3165be8f3ae156f49d3ceae2fb178f892