varnish: parameterize fe cache mem reservation
The size of the main varnish frontend cache memory pool is set by
$fe_mem_gb here, and should be a large fraction of total memory.
More is always better, so long as we don't OOM the host given all
the other stuff running on it.

Historically, we were contending with a couple of issues on this
front:

1. Legacy nodes at various points in the past had as little as 96G
   of total memory, which we've ramped up over time.
2. We've historically had some issues with spikes in varnish's
   transient storage allocations, which necessitated leaving an
   inordinate amount of headroom Just In Case.

Today, even the oldest legacy nodes all have a minimum of 384G of
total memory, and the newest ones arriving this quarter are now
shipping with 512G.  Additionally, we don't seem to have transient
storage memory spike problems anymore, thanks to a number of
investigations and fixes on that front.

For all of these reasons, it's time to try upping the frontend
memory allocation a bit.  Even on the 384G nodes there's too much
(~100GB or so) memory going persistently-unused to justify as
merely headroom, I think.

This patch keeps the legacy calculation `((total_mem - 100GB) *
0.7)` as the default, but allows a new optional per-host hieradata
override called `$fe_mem_gb_reserved`, which if defined causes the
calculation to be `total_mem - fe_mem_gb_reserved`.  We can use
this to experiment with a range of values on the two different
host classes over time and find a new optimal value for each
config and/or cluster.

Change-Id: I84b1547f351f09fcd5da9c2cce5d9b2a71be82e0