doc: add parameters for gitlab_runner and contint hosts, allow them
Allow gitlab_runner hosts to rsync files to doc* servers.

Use the interpolation function (lookup in Hiera itself) so
that we don't have to move host lists to common.yaml or use
aliases and also don't repeat ourselves.

And we can add the gitlab_runner hosts as requested and at the same time
get the list of contint hosts, that are already allowed, from
an existing place in Hiera instead of repeating it here in code.

There should not be hardcoded host names.

Unless lint/CI complains then this should be the easiest way to fix
the linked ticket.

Bug: T321629
Hosts: P:doc
Change-Id: I07d779d2ddb6fd0723884744fd3c0ce85d105b64