Move kubelet configuration to file
The following hiera keys have only one values for all clusters so I
removed the definition from hiera and set a proper default:
* profile::kubernetes::infra_pod: docker-registry.discovery.wmnet/pause
* profile::kubernetes::node::kubelet_cluster_domain: cluster.local

Also --pod-infra-container-image is not really an optional flag in our
infrastructure as the default will simply not work (k8s.gcr.io/pause).

The kubelet port could not be configured via the node role anyways, so I
removed the parameter.

Remove the listen_address from kubelet resouce as it is the default.

Bug T300499

Change-Id: Ieb905505b720a4629278cca6147cbdac70418119