devtools: set profile::gitlab::runner::registration_token: private
This is set in Hiera in Horizon but commented out in the repo.

The goal was to avoid a situation where we have mixed settings in
Horizon and repo hiera where it's unclear which is used.

(Except for spontaenous fixing or short-term testing).

Activating it here to then remove it from the web UI.

Change-Id: Ibd351128b2a18147e7042b463e06b9959898b874