prometheus: Rename ats_ metrics to trafficserver_
The metric names weren't consistent with existing ones. All the previous
metrics are named using the prefix trafficserver so the new metrics
should follow suit.

Fixes an issue that triggered a revert after merging
I06388a6a8731a884a703f56802cbb650b1b2db2b. In that CR
prometheus::node_trafficserver_config was defined as a class
and it should be a define

Bug: T292815
Change-Id: Ia67b3d2d772d3fe73cb4b291856abdbb7f2c89b0