dumps/distribution: move hardcoded host names to parameters
The goal is to get the hard-coded host names out of the puppet
code and into parameters and Hiera.

Add a parameter for the stat dump host and one for the phab dump host.

Also partially follow-up to T322221 though unrelated to the bug there.

And in preparation for the upcoming Phabricator server switch
where we will have to update the phab1001 host name soon.

Bug: T280597
Bug: T322221
Change-Id: I570768bb03df32f954aced32943a8af123e0b441