phabricator: remove phab2001 from the list of phab servers
This is similar to I824e2acc9993c but different.

It removes the host from ssh access between it and other
phab servers.

It also removes it from rsyncd and ferm rules
but for dumps as opposed to rsync for phab data.

See compiler output links on both changes for details.

Bug: T322250
Change-Id: I1f12ea6936eada2b4184820758e0743eab47db6f