query_service: make blazegraph exporter sleep before starting
In Ic2c14347f808de7b82ee8d304e64586cdb5f4612,
we added a dependency of the exporter on blazegraph itself.

In order to achieve our intent of not having the exporter
logspam due to not being able to talk to blazegraph,
we're adding a 30 second sleep at the beginning of the exporter's
startup (otherwise the exporter starts too fast).

Hosts: wcqs2003.codfw.wmnet, wdqs1004.eqiad.wmnet
Bug: T322037
Change-Id: I46591e010c77b7db0319927738f8c28396857aab