growthexperiments.pp: Run updateMetrics.php daily
The Growth team needs to monitor a couple of metrics
that are hard to calculate in an event-driven way
(because they depend on number of new users registered,
or because there are many events that can change
the metric value).

For those metrics, we decided to calculate them daily
via a maintenance script, stored in
extensions/GrowthExperiments/maintenance/updateMetrics.php.
The script makes necessary queries and sends the calculated
value to the statsd server.

The --verbose flag is there to log the calculated values
via syslog as well (this can be used to verify the Grafana
dashboard provides the expected numbers).

Bug: T318684
Depends-On: Ie6c43d1877c5e06750487e982d00f091f00f4592
Hosts: mwmaint1002.eqiad.wmnet
Change-Id: I242715d8faaabc4fde94ded899d7a2659d4f7a54