gitlab_runner: run cleanup of docker cache twice daily
This change increases the clear_interval of docker cache to
every 12h on WMCS Shared Runners. Shared Runners have a 40GB disk
for /var/lib/docker currently which is not enough for holding the
cache of a full day (24h cleanup interval).
Long term bigger disk or other cache storage is needed, as this change
reduces CI performance.

Bug: T310593
Change-Id: Ie28a8703d0d8f715a1809c061ec0e5aef9f92560