package_builder: remove deprecated Varnish6 hooks
The current varnish pbuilder hooks assume a varnish install from the
varnish6 component which has been deprecated and is now instead in main.
Additionally, it seems like these hooks were not installed in
package_builder::pbuilder_hook and therefore were not available on the
build host.

The D01apt.wikimedia.org.erb file and WIKIMEDIA=yes should cover the
varnish pbuilder apt pinning,

	Package: *
	Pin: release o=Wikimedia
	Pin-Priority: 1001

... setting a higher priority for apt.wikimedia.org and our version of
varnish than the version in bullseye. So the current hooks are not being
installed and are also no longer required and this commit removes them.

Bug: T321309
Change-Id: Ifc118d0dc4cad78bae3b2f670272b371365eea8d